some refactors and fixes

+++ src/Service/LessonService.php
@@ -26,9 +26,9 @@ class LessonService
         return $this->lessonMapper->toDTO($lesson);
     }
 
-    public function toEntity (LessonDTO $lessonDTO, Lesson $lesson = null) : Lesson
+    public function toEntity (LessonDTO $lessonDTO) : Lesson
     {
-        return $this->lessonMapper->toEntity($lessonDTO, $lesson);
+        return $this->lessonMapper->toEntity($lessonDTO);
     }
 
     public function getAll() : array
@@ -40,10 +40,10 @@ class LessonService
     public function getLessonById(int $id): ?LessonDTO
     {
         $lesson = $this->lessonRepository->find($id);
-        if ($lesson) {
-            return $this->toDTO($lesson);
+        if (!$lesson) {
+            throw new \RuntimeException("Lesson with ID $id not found.");
         }
-        return null;
+        return $this->toDTO($lesson);
     }
 
     public function getStudentAvailableLessonsBySemester($semester): ?array
@@ -55,20 +55,31 @@ class LessonService
         return null;
     }
     
-    public function save(LessonDTO $lessonDTO, int $id = null) : LessonDTO
+    public function save(LessonDTO $lessonDTO) : LessonDTO
     {
-        $lesson = ($id) ? $this->lessonRepository->find($id) : new Lesson();
-        $lesson = $this->toEntity($lessonDTO, $lesson);
-        $lesson->setUpdatedAt(new \DateTimeImmutable());
-        $this->em->persist($lesson);
-        $this->em->flush();
-        return $this->toDTO($lesson);
+        try {
+            $lesson = $this->toEntity($lessonDTO);
+            $lesson->setUpdatedAt(new \DateTimeImmutable());
+            $this->em->persist($lesson);
+            $this->em->flush();
+            return $this->toDTO($lesson);
+        } catch (\Exception $e) {
+            throw new \RuntimeException("Failed to save lesson: " . $e->getMessage());
+        }
     }
 
     public function delete(int $id) : void
     {
-        $lesson = $this->lessonRepository->find($id);
-        $this->em->remove($lesson);
-        $this->em->flush();
+         $lesson = $this->lessonRepository->find($id);
+        if (!$lesson) {
+            throw new \RuntimeException("Lesson with ID $id not found.");
+        }
+
+        try {
+            $this->em->remove($lesson);
+            $this->em->flush();
+        } catch (\Exception $e) {
+            throw new \RuntimeException("Failed to delete lesson: " . $e->getMessage());
+        }
     }
 }
